+++ src/main/java/kitchenpos/menu/domain/Menu.java
@@ -1,4 +1,4 @@
-package kitchenpos.menu.entity;
+package kitchenpos.menu.domain;
 
 import java.math.BigDecimal;
 import java.util.List;
