+++ src/main/java/kitchenpos/order/domain/OrderDao.java
@@ -1,6 +1,6 @@
-package kitchenpos.order.model;
+package kitchenpos.order.domain;
 
-import kitchenpos.order.entity.Order;
+import kitchenpos.order.domain.Order;
 
 import java.util.List;
 import java.util.Optional;
