+++ src/main/java/kitchenpos/order/domain/OrderStatus.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.entity;
+package kitchenpos.order.domain;
 
 public enum OrderStatus {
     COOKING, MEAL, COMPLETION
