+++ src/main/java/kitchenpos/order/domain/TableGroup.java
@@ -1,4 +1,4 @@
-package kitchenpos.order.entity;
+package kitchenpos.order.domain;
 
 import java.time.LocalDateTime;
 import java.util.List;
