+++ src/main/java/kitchenpos/order/infra/JdbcTemplateOrderLineItemDao.java
@@ -1,6 +1,7 @@
-package kitchenpos.menu.model;
+package kitchenpos.order.infra;
 
-import kitchenpos.menu.entity.Menu;
+import kitchenpos.order.domain.OrderLineItem;
+import kitchenpos.order.domain.OrderLineItemDao;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
@@ -16,14 +17,14 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public class JdbcTemplateMenuDao implements MenuDao {
-    private static final String TABLE_NAME = "menu";
-    private static final String KEY_COLUMN_NAME = "id";
+public class JdbcTemplateOrderLineItemDao implements OrderLineItemDao {
+    private static final String TABLE_NAME = "order_line_item";
+    private static final String KEY_COLUMN_NAME = "seq";
 
     private final NamedParameterJdbcTemplate jdbcTemplate;
     private final SimpleJdbcInsert jdbcInsert;
 
-    public JdbcTemplateMenuDao(final DataSource dataSource) {
+    public JdbcTemplateOrderLineItemDao(final DataSource dataSource) {
         jdbcTemplate = new NamedParameterJdbcTemplate(dataSource);
         jdbcInsert = new SimpleJdbcInsert(dataSource)
                 .withTableName(TABLE_NAME)
@@ -32,14 +33,14 @@ public class JdbcTemplateMenuDao implements MenuDao {
     }
 
     @Override
-    public Menu save(final Menu entity) {
+    public OrderLineItem save(final OrderLineItem entity) {
         final SqlParameterSource parameters = new BeanPropertySqlParameterSource(entity);
         final Number key = jdbcInsert.executeAndReturnKey(parameters);
         return select(key.longValue());
     }
 
     @Override
-    public Optional<Menu> findById(final Long id) {
+    public Optional<OrderLineItem> findById(final Long id) {
         try {
             return Optional.of(select(id));
         } catch (final EmptyResultDataAccessException e) {
@@ -48,32 +49,32 @@ public class JdbcTemplateMenuDao implements MenuDao {
     }
 
     @Override
-    public List<Menu> findAll() {
-        final String sql = "SELECT id, name, price, menu_group_id FROM menu ";
+    public List<OrderLineItem> findAll() {
+        final String sql = "SELECT seq, order_id, menu_id, quantity FROM order_line_item";
         return jdbcTemplate.query(sql, (resultSet, rowNumber) -> toEntity(resultSet));
     }
 
     @Override
-    public long countByIdIn(final List<Long> ids) {
-        final String sql = "SELECT COUNT(*) FROM menu WHERE id IN (:ids)";
+    public List<OrderLineItem> findAllByOrderId(final Long orderId) {
+        final String sql = "SELECT seq, order_id, menu_id, quantity FROM order_line_item WHERE order_id = (:orderId)";
         final SqlParameterSource parameters = new MapSqlParameterSource()
-                .addValue("ids", ids);
-        return jdbcTemplate.queryForObject(sql, parameters, Long.class);
+                .addValue("orderId", orderId);
+        return jdbcTemplate.query(sql, parameters, (resultSet, rowNumber) -> toEntity(resultSet));
     }
 
-    private Menu select(final Long id) {
-        final String sql = "SELECT id, name, price, menu_group_id FROM menu WHERE id = (:id)";
+    private OrderLineItem select(final Long id) {
+        final String sql = "SELECT seq, order_id, menu_id, quantity FROM order_line_item WHERE seq = (:seq)";
         final SqlParameterSource parameters = new MapSqlParameterSource()
-                .addValue("id", id);
+                .addValue("seq", id);
         return jdbcTemplate.queryForObject(sql, parameters, (resultSet, rowNumber) -> toEntity(resultSet));
     }
 
-    private Menu toEntity(final ResultSet resultSet) throws SQLException {
-        final Menu entity = new Menu();
-        entity.setId(resultSet.getLong("id"));
-        entity.setName(resultSet.getString("name"));
-        entity.setPrice(resultSet.getBigDecimal("price"));
-        entity.setMenuGroupId(resultSet.getLong("menu_group_id"));
+    private OrderLineItem toEntity(final ResultSet resultSet) throws SQLException {
+        final OrderLineItem entity = new OrderLineItem();
+        entity.setSeq(resultSet.getLong(KEY_COLUMN_NAME));
+        entity.setOrderId(resultSet.getLong("order_id"));
+        entity.setMenuId(resultSet.getLong("menu_id"));
+        entity.setQuantity(resultSet.getLong("quantity"));
         return entity;
     }
 }
