+++ src/main/java/kitchenpos/order/infra/JdbcTemplateTableGroupDao.java
@@ -1,6 +1,7 @@
-package kitchenpos.order.model;
+package kitchenpos.order.infra;
 
-import kitchenpos.order.entity.TableGroup;
+import kitchenpos.order.domain.TableGroup;
+import kitchenpos.order.domain.TableGroupDao;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
