+++ src/main/java/kitchenpos/order/service/TableBo.java
@@ -1,9 +1,9 @@
 package kitchenpos.order.service;
 
-import kitchenpos.order.model.OrderDao;
-import kitchenpos.order.model.OrderTableDao;
-import kitchenpos.order.entity.OrderStatus;
-import kitchenpos.order.entity.OrderTable;
+import kitchenpos.order.domain.OrderDao;
+import kitchenpos.order.domain.OrderTableDao;
+import kitchenpos.order.domain.OrderStatus;
+import kitchenpos.order.domain.OrderTable;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
