+++ src/main/java/kitchenpos/product/domain/ProductDao.java
@@ -1,6 +1,6 @@
-package kitchenpos.product.model;
+package kitchenpos.product.domain;
 
-import kitchenpos.product.entity.Product;
+import kitchenpos.product.domain.Product;
 
 import java.util.List;
 import java.util.Optional;
