+++ src/test/java/kitchenpos/bo/ProductBoTest.java
@@ -1,7 +1,7 @@
 package kitchenpos.bo;
 
 import kitchenpos.product.service.ProductBo;
-import kitchenpos.product.model.ProductDao;
+import kitchenpos.product.domain.ProductDao;
 import kitchenpos.model.Product;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
