+++ src/test/java/kitchenpos/controller/MenuGroupRestControllerTest.java
@@ -2,7 +2,7 @@ package kitchenpos.controller;
 
 import kitchenpos.menu.service.MenuGroupBo;
 import kitchenpos.menu.controller.MenuGroupRestController;
-import kitchenpos.menu.entity.MenuGroup;
+import kitchenpos.menu.domain.MenuGroup;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.autoconfigure.web.servlet.HttpEncodingAutoConfiguration;
